Software_Testing_Project-Manual_Testing-01OpenCart
The scope of this project is limited to the testing of the features described in the succeeding sections of this document. On-functional testing like stress, performance is beyond scope of this project. 
OpenCart is a free open-source e-commerce platform for online merchants. OpenCart provides a professional and reliable foundation from which to build a successful online store. This foundation appeals to a wide variety of users;