Manual_Testing_Project_02_guru99
The scope of this project is limited to the testing of the features described in the succeeding sections of this document. On-functional testing like stress, performance is beyond scope of this project. Automation testing is beyond the scope. Functional testing & external interfaces are in scope and need to be tested The banking site will be only compatible with Chrome version 27 and above